Renaming (and moving) config parameters component to more generic name
